refactor: move mobile helpers
