--- tooling/cli/src/helpers/mobile/android/mod.rs
@@ -1,1 +1,0 @@
-pub mod project;
