--- tooling/cli/src/helpers/mobile/ios/mod.rs
@@ -1,1 +1,0 @@
-pub mod project;
