--- tooling/cli/src/helpers/mobile/mod.rs
@@ -1,4 +1,0 @@
-pub mod android;
-pub mod init;
-#[cfg(target_os = "macos")]
-pub mod ios;
