+++ tooling/cli/src/mobile/ios/project.rs
@@ -1,3 +1,7 @@
+// Copyright 2019-2021 Tauri Programme within The Commons Conservancy
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: MIT
+
 use crate::helpers::template;
 use cargo_mobile::{
   apple::{
